backend/requirements.txt
# FastAPI核心
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# 数据库
sqlalchemy==2.0.23
aiomysql==0.2.0
pymysql==1.1.0
alembic==1.12.1

# Redis和Celery
redis==5.0.1
celery==5.3.4

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP客户端
httpx==0.25.2
aiohttp==3.9.1

# AI服务SDK
openai==1.3.7
anthropic==0.7.7

# 工具库
python-dotenv==1.0.0
pyyaml==6.0.1

# 日志
loguru==0.7.2

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
